backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
Werkzeug==3.0.1
